New plugin repo cloned using new GitRepo class. Readme and info displayed in wide modal dialog.
